+++ Changes.md
@@ -9,6 +9,7 @@ Improvements
 - NodeEditor :
   - Added BoxPlug items to the plug creation menus.
   - Added more Array items to the plug creation menus.
+- CustomAttributes : Added "From Selected" and "From Affected" submenus to the plug creation menu.
 
 Fixes
 -----